requirements.txt
streamlit==1.36.0
sentence-transformers==2.7.0
groq>=0.4.0
faiss-cpu==1.8.0.post1
PyYAML==6.0.1
pydantic==2.8.2
python-dotenv==1.0.1
numpy
